requirements.txt
python-telegram-bot==13.14
wikipedia==1.4.0

